@@ -13,11 +13,11 @@
   ],
 
   "require": {
-    "php": ">=5.3.0"
+    "php": ">=7.4.0"
   },
 
   "require-dev": {
-    "phpunit/phpunit": "~4.6.0"
+    "phpunit/phpunit": ">9.0.0"
   },
 
   "archive": {
