@@ -16,7 +16,7 @@ use org\camunda\php\sdk\service\ProcessInstanceService;
 
 include('../../vendor/autoload.php');
 
-class ProcessDefinitionTest extends \PHPUnit_Framework_TestCase {
+class ProcessDefinitionTest extends \PHPUnit\Framework\TestCase {
   protected static $restApi;
   protected static $pds;
 
