@@ -12,7 +12,7 @@ use org\camunda\php\sdk\service\ProcessEngineService;
 
 include('../../vendor/autoload.php');
 
-class ProcessEngineServiceTest extends \PHPUnit_Framework_TestCase {
+class ProcessEngineServiceTest extends \PHPUnit\Framework\TestCase {
   protected static $restApi;
   protected static $pes;
 
